requirements.txt
Pillow>=10.0.0
pillow-heif>=0.16.0
openai>=1.40.0
